fs/cas: move CLI command handlers to fs/cas/cli/module.f.ts
